+++ felix/deps.txt
@@ -367,7 +367,6 @@ local:libcalico-go/lib/selector
 local:libcalico-go/lib/selector/parser
 local:libcalico-go/lib/selector/tokenizer
 local:libcalico-go/lib/set
-local:libcalico-go/lib/validator/v1
 local:libcalico-go/lib/validator/v3
 local:libcalico-go/lib/watch
 local:libcalico-go/lib/winutils
@@ -378,3 +377,4 @@ local:typha/pkg/discovery
 local:typha/pkg/syncclient
 local:typha/pkg/syncproto
 local:typha/pkg/tlsutils
+local:typha/pkg/validator/v1
